dtype for cnn

+++ flappy_env.py
@@ -45,7 +45,7 @@ class FlappyEnv(gym.Env):
 		# self.observation_space = gym.spaces.Box(-np.inf, np.inf,
 		# 								shape=(2,),
 		# 								dtype=np.float32)
-		self.observation_space = spaces.Box(low = 0, high = 255, shape = (SCREENHEIGHT, SCREENWIDTH, 3))
+		self.observation_space = spaces.Box(low = 0, high = 255, shape = (SCREENHEIGHT, SCREENWIDTH, 3), dtype=np.uint8)
 
 		pygame.init()
 		self.FPSCLOCK = pygame.time.Clock()
